requirements.txt
Flask
openai-whisper
pyngrok
requests